+++ ontimize/components/table/extensions/header/table-menu/o-table-menu.component.ts
@@ -294,6 +294,7 @@ export class OTableMenuComponent implements OnInit, AfterViewInit, OnDestroy {
     // get column's attr whose renderer is OTableCellRendererImageComponent
     let colsNotIncluded: string[] = tableOptions.columns.filter(c => void 0 !== c.renderer && c.renderer instanceof OTableCellRendererImageComponent).map(c => c.attr);
     colsNotIncluded.push(OTableComponent.NAME_COLUMN_SELECT);
+    colsNotIncluded.push(Codes.NAME_COLUMN_EXPANDABLE);
 
     // Table data/filters
     switch (this.table.exportMode) {